+++ scripts/src/lib.lua
@@ -41,6 +41,7 @@ project "utils"
 		MAME_DIR .. "src/lib/util/chdcd.h",
 		MAME_DIR .. "src/lib/util/chdcodec.cpp",
 		MAME_DIR .. "src/lib/util/chdcodec.h",
+		MAME_DIR .. "src/lib/util/contiguous_sequence_wrapper.h",
 		MAME_DIR .. "src/lib/util/corealloc.h",
 		MAME_DIR .. "src/lib/util/corefile.cpp",
 		MAME_DIR .. "src/lib/util/corefile.h",
@@ -93,7 +94,6 @@ project "utils"
 		MAME_DIR .. "src/lib/util/vbiparse.h",
 		MAME_DIR .. "src/lib/util/vecstream.cpp",
 		MAME_DIR .. "src/lib/util/vecstream.h",
-		MAME_DIR .. "src/lib/util/view.h",
 		MAME_DIR .. "src/lib/util/wavwrite.cpp",
 		MAME_DIR .. "src/lib/util/wavwrite.h",
 		MAME_DIR .. "src/lib/util/xmlfile.cpp",
